Add solution for day 6
Solution in ruby for sorting the customs declaration challenge input by repeating characters (day6_1.rb) and by repeating characters present in all input strings from a single group (day6_2.rb).
Not the most ideal solution as the input needs to be modified to have a newline at the end of the file so that File.readlines('input') is usable.
Either way, an enjoyable challenge and one that shows the power of Ruby's single-line solutions (e.g. concat a string, then seperate it into characters, then group then by a condition, etc).